Fix windows compilation (#274)
* Fix VS 2015 windows compilation: conversion from 'const unsigned int' to 'int' requires a narrowing conversion

* Missing new source file faceDetectorOpenCV.cpp in VS 2015 projects

* Fix VS 2015 windows compilation: conversion from 'const unsigned int' to 'int' requires a narrowing conversion

+++ src/openpose/hand/handExtractor.cpp
@@ -219,8 +219,8 @@ namespace op
                 // HeatMaps: define size
                 if (!mHeatMapTypes.empty())
                 {
-                    mHeatMaps[0].reset({numberPeople, HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
-                    mHeatMaps[1].reset({numberPeople, HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
+                    mHeatMaps[0].reset({numberPeople, (int)HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
+                    mHeatMaps[1].reset({numberPeople, (int)HAND_NUMBER_PARTS, mNetOutputSize.y, mNetOutputSize.x});
                 }
 
                 // // Debugging
